// frontend/src/api-generated/models/FriendCollectionsDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Momocraft Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    UserFriendCollectionDto,
    UserFriendCollectionDtoFromJSON,
    UserFriendCollectionDtoFromJSONTyped,
    UserFriendCollectionDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface FriendCollectionsDto
 */
export interface FriendCollectionsDto {
    /**
     * 
     * @type {Array<UserFriendCollectionDto>}
     * @memberof FriendCollectionsDto
     */
    friendCollections: Array<UserFriendCollectionDto>;
}

export function FriendCollectionsDtoFromJSON(json: any): FriendCollectionsDto {
    return FriendCollectionsDtoFromJSONTyped(json, false);
}

export function FriendCollectionsDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): FriendCollectionsDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'friendCollections': ((json['friendCollections'] as Array<any>).map(UserFriendCollectionDtoFromJSON)),
    };
}

export function FriendCollectionsDtoToJSON(value?: FriendCollectionsDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'friendCollections': ((value.friendCollections as Array<any>).map(UserFriendCollectionDtoToJSON)),
    };
}