// frontend/src/api-generated/models/FriendUserDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Momocraft Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FriendUserDto
 */
export interface FriendUserDto {
    /**
     * 
     * @type {string}
     * @memberof FriendUserDto
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof FriendUserDto
     */
    comment: string;
    /**
     * 
     * @type {string}
     * @memberof FriendUserDto
     */
    avatarUrl?: string;
    /**
     * 
     * @type {number}
     * @memberof FriendUserDto
     */
    approvedAt: number;
}

export function FriendUserDtoFromJSON(json: any): FriendUserDto {
    return FriendUserDtoFromJSONTyped(json, false);
}

export function FriendUserDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): FriendUserDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'userId': json['userId'],
        'comment': json['comment'],
        'avatarUrl': !exists(json, 'avatarUrl') ? undefined : json['avatarUrl'],
        'approvedAt': json['approvedAt'],
    };
}

export function FriendUserDtoToJSON(value?: FriendUserDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'userId': value.userId,
        'comment': value.comment,
        'avatarUrl': value.avatarUrl,
        'approvedAt': value.approvedAt,
    };
}