// frontend/src/api-generated/models/ItemCollectionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Momocraft Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ItemCollectionDto
 */
export interface ItemCollectionDto {
  /**
   *
   * @type {string}
   * @memberof ItemCollectionDto
   */
  itemId: string;
  /**
   *
   * @type {number}
   * @memberof ItemCollectionDto
   */
  state: number;
}

export function ItemCollectionDtoFromJSON(json: any): ItemCollectionDto {
  return ItemCollectionDtoFromJSONTyped(json, false);
}

export function ItemCollectionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ItemCollectionDto {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    itemId: json['itemId'],
    state: json['state'],
  };
}

export function ItemCollectionDtoToJSON(value?: ItemCollectionDto | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    itemId: value.itemId,
    state: value.state,
  };
}