// frontend/src/api-generated/models/UserFriendCollectionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Momocraft Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
  FriendUserDto,
  FriendUserDtoFromJSON,
  FriendUserDtoFromJSONTyped,
  FriendUserDtoToJSON,
  ItemCollectionDto,
  ItemCollectionDtoFromJSON,
  ItemCollectionDtoFromJSONTyped,
  ItemCollectionDtoToJSON,
} from './';

/**
 *
 * @export
 * @interface UserFriendCollectionDto
 */
export interface UserFriendCollectionDto {
  /**
   *
   * @type {FriendUserDto}
   * @memberof UserFriendCollectionDto
   */
  friend: FriendUserDto;
  /**
   *
   * @type {Array<ItemCollectionDto>}
   * @memberof UserFriendCollectionDto
   */
  friendCollections: Array<ItemCollectionDto>;
}

export function UserFriendCollectionDtoFromJSON(json: any): UserFriendCollectionDto {
  return UserFriendCollectionDtoFromJSONTyped(json, false);
}

export function UserFriendCollectionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserFriendCollectionDto {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    friend: FriendUserDtoFromJSON(json['friend']),
    friendCollections: (json['friendCollections'] as Array<any>).map(ItemCollectionDtoFromJSON),
  };
}

export function UserFriendCollectionDtoToJSON(value?: UserFriendCollectionDto | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    friend: FriendUserDtoToJSON(value.friend),
    friendCollections: (value.friendCollections as Array<any>).map(ItemCollectionDtoToJSON),
  };
}