// frontend/src/api-generated/models/FriendListDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Momocraft Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    FriendUserDto,
    FriendUserDtoFromJSON,
    FriendUserDtoFromJSONTyped,
    FriendUserDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface FriendListDto
 */
export interface FriendListDto {
    /**
     * 
     * @type {Array<FriendUserDto>}
     * @memberof FriendListDto
     */
    friends: Array<FriendUserDto>;
}

export function FriendListDtoFromJSON(json: any): FriendListDto {
    return FriendListDtoFromJSONTyped(json, false);
}

export function FriendListDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): FriendListDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'friends': ((json['friends'] as Array<any>).map(FriendUserDtoFromJSON)),
    };
}

export function FriendListDtoToJSON(value?: FriendListDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'friends': ((value.friends as Array<any>).map(FriendUserDtoToJSON)),
    };
}